Notify user to check their email for OTP
